Added working verison of the SAS
There is a couple of more things I want to talk about, but this is a good idea
of the style I want for the SAS workshop. I'm trying to incrementally add new
things so that learning is slow and steady. I am concurrently writing a guide
on how to get SAS working in Jupyter Notebook, that will have to be in PDF
format and be an associated file with the workshop. The instructions for SAS
University Edition (i.e. the free one) is very clear on the SAS site, so I
think I will point people to that for that leg of the tutorial.